🏗️ Architect: [Refactor] audio-system.ts into core and playback modules
Extracts the large `src/audio/audio-system.ts` monolith into specialized modules (`audio-system-core.ts` for WASM setup and types, and `audio-system-playback.ts` for module playback), and uses the original file as a barrel export. Updates roadmaps. No logical changes.
